# docs/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ICoder API",
    "description": "API Documentation for ICoder Project",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/auth/verify/send": {
      "post": {
        "tags": [
          "Authentication Management"
        ],
        "summary": "Send email verification link",
        "description": "Sends a new email containing the verification link to the user's email address.",
        "operationId": "sendVerificationEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendVerificationEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Authentication Management"
        ],
        "summary": "Register a new user",
        "description": "Creates a new user account with the provided details and needs to be verified",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/password/reset": {
      "post": {
        "tags": [
          "Authentication Management"
        ],
        "summary": "Reset user password",
        "description": "Resets the user's password using a valid reset token and a new password.",
        "operationId": "resetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password/forget": {
      "post": {
        "tags": [
          "Authentication Management"
        ],
        "summary": "Initiate password reset process",
        "description": "Sends a password reset link to the user's email address.",
        "operationId": "forgetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication Management"
        ],
        "summary": "User login",
        "description": "Authenticates a user with handle and password and sets authentication/refresh cookies.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/password": {
      "patch": {
        "tags": [
          "Authentication Management"
        ],
        "summary": "Change authenticated user's password",
        "description": "Allows an authenticated user to change their password by providing the old and new passwords.",
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/summary/{userId}": {
      "get": {
        "tags": [
          "AI Summary Management"
        ],
        "summary": "Get AI performance summary",
        "description": "Aggregates user submissions and returns an AI-generated coaching summary",
        "operationId": "getSummary",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summary/{userId}/stats-only": {
      "get": {
        "tags": [
          "AI Summary Management"
        ],
        "summary": "Get raw stats only (no AI)",
        "description": "Returns the aggregated stats without calling the AI — useful for testing",
        "operationId": "getStatsOnly",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/verify": {
      "get": {
        "tags": [
          "Authentication Management"
        ],
        "summary": "Verify user email",
        "description": "Confirms the user's email address using a verification token sent via email.",
        "operationId": "verifyEmail",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "SendVerificationEmailRequest": {
        "required": [
          "handle"
        ],
        "type": "object",
        "properties": {
          "handle": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "required": [
          "email",
          "handle",
          "nickname",
          "password",
          "password_confirmation"
        ],
        "type": "object",
        "properties": {
          "handle": {
            "maxLength": 20,
            "minLength": 3,
            "pattern": "^[a-z][a-z0-9_]{2,19}$",
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "maxLength": 20,
            "minLength": 8,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]+$",
            "type": "string"
          },
          "password_confirmation": {
            "type": "string"
          },
          "school": {
            "type": "string"
          }
        }
      },
      "ResetPasswordRequest": {
        "required": [
          "confirmation_password",
          "new_password",
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "new_password": {
            "maxLength": 20,
            "minLength": 8,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]+$",
            "type": "string"
          },
          "confirmation_password": {
            "type": "string"
          }
        }
      },
      "ForgetPasswordRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "required": [
          "handle",
          "password"
        ],
        "type": "object",
        "properties": {
          "handle": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          }
        }
      },
      "ChangePasswordRequest": {
        "required": [
          "current_password",
          "new_password",
          "password_confirmation"
        ],
        "type": "object",
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "maxLength": 20,
            "minLength": 8,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]+$",
            "type": "string"
          },
          "password_confirmation": {
            "type": "string"
          }
        }
      },
...
      "UserStatsDto": {
        "type": "object",
        "properties": {
          "totalSubmissions": {
            "type": "integer",
            "format": "int64"
          },
          "overallAcRate": {
            "type": "number",
            "format": "double"
          },
          "recentAcRate": {
            "type": "number",
            "format": "double"
          },
          "tleCount": {
            "type": "integer",
            "format": "int32"
          },
          "rteCount": {
            "type": "integer",
            "format": "int32"
          },
          "mleCount": {
            "type": "integer",
            "format": "int32"
          },
          "compilationErrorCount": {
            "type": "integer",
            "format": "int32"
          },
          "strengths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "weaknesses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mostUsedLanguage": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}